Mitigate collapsing visual bug
Bug was happening where split second collapses of the mood icon container had a height of zero shifting everything back and forth

Introduced a min-height that is defined by the height of the svg for scalability to desktop

This mitigates the bug but does not fix it entirely. Some pixels still shift up, but it is not nearly as visually jarring.

This pattern is unpredictable on Chrome, but occurs on first click for every button, then functions normally after on Firefox.