Nav Responsiveness
Updated layout of nav to change above 800px.

To properly apply it to every page, this nav needs to implemented on a grid. This is applied to layout with consideration for whether or not this screen should have a nav or not. This has been tested on every page, and works so far.

Nav takes up 324px, rest of screen fills to 100% of the available space.