Font size + Streak
Font size for the heading is calculated by a function seeing the windows inner width then applying the proper h1 or h2 font-size. This does not automatically change when screen sizes are adjusted and needs a refresh.

Streak given responsive home page design

Nav tweaked to better account for the
 8px stroke