Exercise Info Block Fixed
This problem was caused by the way exercise model scaled up with different page sizes. When shrinking from a larger page size, the screen would not recognize that it had adapted and the exercise model would end up with a larger height than it started with.

This is fixed by adding an aspect ratio and minimum height with some parameters to make sure it runs correctly.